Provide SegmentContextFactory from Tracer
