+++ cpp2sky/segment_context.h
@@ -271,30 +271,4 @@ class SegmentContext {
 
 using SegmentContextPtr = std::shared_ptr<SegmentContext>;
 
-class SegmentContextFactory {
- public:
-  virtual ~SegmentContextFactory() = default;
-
-  /**
-   * Create segment context that doesn't have propagated info.
-   */
-  virtual SegmentContextPtr create() = 0;
-
-  /**
-   * Create segment context with propagated span context.
-   */
-  virtual SegmentContextPtr create(SpanContextPtr span_context) = 0;
-
-  /**
-   * Create segment context with propagated span context and extensions.
-   */
-  virtual SegmentContextPtr create(
-      SpanContextPtr span_context,
-      SpanContextExtensionPtr ext_span_context) = 0;
-};
-
-using SegmentContextFactoryPtr = std::unique_ptr<SegmentContextFactory>;
-
-SegmentContextFactoryPtr createSegmentContextFactory(const TracerConfig& cfg);
-
 }  // namespace cpp2sky
