+++ example/sample_client.cc
@@ -34,12 +34,11 @@ void init() {
 int main() {
   init();
 
-  SegmentContextFactoryPtr factory = createSegmentContextFactory(config);
-
   // 1. Create tracer object to send span data to OAP.
   auto tracer = createInsecureGrpcTracer(config);
+
   // 2. Create segment context
-  auto current_segment = factory->create();
+  auto current_segment = tracer->newSegment();
 
   // 3. Initialize span data to track root workload on current service.
   auto current_span = current_segment->createCurrentSegmentRootSpan();
