+++ test/e2e/consumer.cc
@@ -44,13 +44,12 @@ int main() {
   httplib::Server svr;
   auto tracer = createInsecureGrpcTracer(config);
 
-  SegmentContextFactoryPtr factory = createSegmentContextFactory(config);
-
   svr.Get("/pong", [&](const httplib::Request& req, httplib::Response& res) {
     if (req.has_header(kPropagationHeader.data())) {
       auto parent = req.get_header_value(kPropagationHeader.data());
       auto parent_span = createSpanContext(parent);
-      auto current_segment = factory->create(parent_span);
+
+      auto current_segment = tracer->newSegment(parent_span);
       handlePong(tracer.get(), current_segment.get(), req, res);
       tracer->sendSegment(std::move(current_segment));
     }
